requirements.txt
# Core Dependencies
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# MLflow for experiment tracking and model registry
mlflow>=1.20.0

# Feast for feature store
feast>=0.15.0

# Apache Airflow for pipeline orchestration
apache-airflow>=2.2.0
connexion>=2.14.0

# FastAPI for model serving
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Database connectors
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Data validation
great-expectations>=0.13.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Utilities
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Monitoring
prometheus-client>=0.11.0
prometheus-async>=0.11.0
